ply loader

+++ include/pbase/types/cloud.h
@@ -1,7 +1,6 @@
 #ifndef PBASE_TYPES_CLOUD_H
 #define PBASE_TYPES_CLOUD_H
 
-#include <stddef.h> // size_t
 #include <stdbool.h>
 #include "pbase/mathc/types/float.h"
 #include "indices.h"
@@ -12,7 +11,7 @@
  */
 typedef struct {
     vec4 *restrict data;
-    size_t size;
+    int size;
 } pCloud;
 
 
@@ -27,13 +26,13 @@ static pCloud p_cloud_new_invalid() {
 }
 
 /** Creates a new Cloud structure with empty data (malloc) */
-pCloud p_cloud_new_empty(size_t size);
+pCloud p_cloud_new_empty(int size);
 
 /** Creates a new Cloud structure with all zeros (calloc) */
-pCloud p_cloud_new_zeros(size_t size);
+pCloud p_cloud_new_zeros(int size);
 
 /** Creates a new Cloud structure with all zeros but .w=1 */
-pCloud p_cloud_new_zeros_1(size_t size);
+pCloud p_cloud_new_zeros_1(int size);
 
 /** Frees a Cloud structure */
 void p_cloud_kill(pCloud *self);
