+++ include/pbase/types/matrix.h
@@ -1,7 +1,6 @@
 #ifndef PBASE_TYPES_MATRIX_H
 #define PBASE_TYPES_MATRIX_H
 
-#include <stddef.h> // size_t
 #include <stdbool.h>
 
 /**
@@ -9,13 +8,13 @@
 */
 typedef struct {
     float *restrict data;
-    size_t cols, rows;
+    int cols, rows;
 } pMatrix;
 
 
 /** Returns true if the matrix is in a valid state */
 static bool p_matrix_valid(pMatrix self) {
-    return self.data  && self.cols>0&& self.rows>0;
+    return self.data  && self.cols>0 && self.rows>0;
 }
 
 /** Returns an invalid Matrix */
@@ -24,10 +23,10 @@ static pMatrix p_matrix_new_invalid() {
 }
 
 /** Creates a new Matrix structure with empty data (malloc) */
-pMatrix p_matrix_new_empty(size_t cols, size_t rows);
+pMatrix p_matrix_new_empty(int cols, int rows);
 
 /** Creates a new Matrix structure with all zeros (calloc) */
-pMatrix p_matrix_new_zeros(size_t cols, size_t rows);
+pMatrix p_matrix_new_zeros(int cols, int rows);
 
 /** Frees a Matrix structure */
 void p_matrix_kill(pMatrix *self);
