+++ src/types/matrix.c
@@ -6,7 +6,7 @@
 
 
 /** Creates a new Matrix structure with empty data (malloc) */
-pMatrix p_matrix_new_empty(size_t cols, size_t rows) {
+pMatrix p_matrix_new_empty(int cols, int rows) {
     pMatrix self = {0};
     self.data = p_rhc_malloc_raising(cols * rows * sizeof *self.data);
     self.cols = cols;
@@ -15,7 +15,7 @@ pMatrix p_matrix_new_empty(size_t cols, size_t rows) {
 }
 
 /** Creates a new Matrix structure with all zeros (calloc) */
-pMatrix p_matrix_new_zeros(size_t cols, size_t rows) {
+pMatrix p_matrix_new_zeros(int cols, int rows) {
     pMatrix self = p_matrix_new_empty(cols, rows);
     memset(self.data, 0, self.cols * self.rows * sizeof *self.data);
     return self;
@@ -37,8 +37,8 @@ void p_matrix_print(pMatrix self) {
         return;
     }
     puts("pc_matrix_print:");
-    for(size_t r = 0; r < self.rows; r++) {
-        for(size_t c = 0; c < self.cols; c++) {
+    for(int r = 0; r < self.rows; r++) {
+        for(int c = 0; c < self.cols; c++) {
             printf("%f ", self.data[r*self.cols + c]);
         }
         printf("\n");
