+++ src/types/meshindices.c
@@ -7,22 +7,22 @@
 #include "pbase/types/meshindices.h"
 
 
-pMeshIndices p_mesh_indices_new_empty(size_t size) {
+pMeshIndices p_mesh_indices_new_empty(int size) {
     pMeshIndices self = {0};
     self.data = p_rhc_malloc_raising(size * sizeof *self.data);
     self.size = size;
     return self;
 }
 
-pMeshIndices p_mesh_indices_new_zeros(size_t size) {
+pMeshIndices p_mesh_indices_new_zeros(int size) {
     pMeshIndices self = p_mesh_indices_new_empty(size);
     memset(self.data, 0, self.size * sizeof *self.data);
     return self;
 }
 
-pMeshIndices p_mesh_indices_new_count_up(size_t size) {
+pMeshIndices p_mesh_indices_new_count_up(int size) {
     pMeshIndices self = p_mesh_indices_new_empty(size);
-    for(size_t i=0; i<self.size; i++) {
+    for(int i=0; i<self.size; i++) {
         self.data[i].v0 = i*3;
         self.data[i].v1 = i*3 + 1;
         self.data[i].v2 = i*3 + 2;
@@ -44,13 +44,13 @@ void p_mesh_indices_print(pMeshIndices self) {
         return;
     }
     puts("p_mesh_indices_print:");
-    for(size_t i = 0; i < self.size; i++) {
+    for(int i = 0; i < self.size; i++) {
         printf("%d %d %d\n", self.data[i].v0, self.data[i].v1, self.data[i].v2);
     }
 }
 
 void p_mesh_indices_add_offset(pMeshIndices *self, int offset) {
-    for(size_t i=0; i<self->size; i++) {
+    for(int i=0; i<self->size; i++) {
         for(int abc=0; abc<3; abc++)
             self->data[i].v[abc] += offset;
     }
@@ -61,7 +61,7 @@ pMeshIndices p_mesh_indices_concatenate(pMeshIndices a, pMeshIndices b) {
 }
 
 pMeshIndices p_mesh_indices_concatenate_v(const pMeshIndices *mesh_indices_list, int n) {
-    size_t size = 0;
+    int size = 0;
     for(int i=0; i<n; i++)
         size += mesh_indices_list[i].size;
     pMeshIndices self = p_mesh_indices_new_empty(size);
@@ -77,7 +77,7 @@ pMeshIndices p_mesh_indices_concatenate_v(const pMeshIndices *mesh_indices_list,
 
 pMeshIndices p_mesh_indices_apply_indices(pMeshIndices self, pIndices indices) {
     pMeshIndices ret = p_mesh_indices_new_empty(indices.size);
-    for(size_t i=0; i<indices.size; i++) {
+    for(int i=0; i<indices.size; i++) {
         int index = isca_mod_positive(indices.data[i], self.size);
         ret.data[i] = self.data[index];
     }
@@ -87,7 +87,7 @@ pMeshIndices p_mesh_indices_apply_indices(pMeshIndices self, pIndices indices) {
 
 ivec3 p_mesh_indices_min(pMeshIndices self) {
     ivec3 min = ivec3_set(INT_MAX);
-    for(size_t i=0; i<self.size; i++) {
+    for(int i=0; i<self.size; i++) {
         for(int abc=0; abc < 3; abc++) {
             if(min.v[abc] > self.data[i].v[abc])
                 min.v[abc] = self.data[i].v[abc];
@@ -98,7 +98,7 @@ ivec3 p_mesh_indices_min(pMeshIndices self) {
 
 ivec3 p_mesh_indices_max(pMeshIndices self) {
     ivec3 max = ivec3_set(INT_MIN);
-    for(size_t i=0; i<self.size; i++) {
+    for(int i=0; i<self.size; i++) {
         for(int abc=0; abc < 3; abc++) {
             if(max.v[abc] < self.data[i].v[abc])
                 max.v[abc] = self.data[i].v[abc];
